Made a workspace for building the static lib
