+++ .gitignore
@@ -1,3 +1,3 @@
-/target
+target/
 /Cargo.lock
 .DS_Store
