+++ Cargo.toml
@@ -1,29 +1,5 @@
-[package]
-name = "laspa"
-version = "0.3.1"
-edition = "2021"
-description = "A simple Lisp-like language built with Rust"
-readme = "README.md"
-repository = "https://github.com/manorajesh/laspa"
-license = "MIT"
-keywords = ["language", "compiler", "interpreter", "programming-language", "parse"]
-categories = ["compilers"]
-exclude = ["examples", ".github"]
-
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-clap = { version = "4.3.22", features = ["derive"] }
-env_logger = "0.10.0"
-indicatif = "0.17.6"
-inkwell = { version = "0.2.0", features = ["llvm16-0"] }
-lazy_static = { version = "1.4.0", default-features = false }
-log = "0.4.20"
-regex = { version = "1.9.3", default-features = false }
-
-[lib]
-name = "laspa"
-
-[[bin]]
-name = "laspa"
+[workspace]
+members = [
+    "laspa-toolchain",
+    "laspa-std"
+]
