+++ laspa-std/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "laspa_std"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[lib]
+crate-type = ["staticlib"]
